Drivetrain

+++ robot/src/main/include/subsystems/Drivetrain.h
@@ -36,7 +36,7 @@ public:
 
 	double GetRoll();
 	double GetPitch();
-	double GetYaw();
+	int GetYaw();
 	int16_t GetX();
 	int16_t GetY();
 	int16_t GetZ();
